+++ packages/ocean-react/src/Select/stories/_shared.tsx
@@ -495,6 +495,368 @@ export const COMMON_SELECT_CSS_CLASSES: CssClass[] = [
   },
 ];
 
+// CSS Classes para SelectAutocomplete
+export const COMMON_SELECT_AUTOCOMPLETE_CSS_CLASSES: CssClass[] = [
+  {
+    name: '.ods-select-autocomplete__root',
+    description: 'Estilos aplicados ao elemento raiz.',
+  },
+  {
+    name: '.ods-select-autocomplete__arrow',
+    description: 'Estilos aplicados ao elemento de seta.',
+  },
+  {
+    name: '.ods-select-autocomplete__arrow--up',
+    description: 'Estilos aplicados à seta quando o popup está aberto.',
+  },
+  {
+    name: '.ods-select-autocomplete__arrow--down',
+    description: 'Estilos aplicados à seta quando o popup está fechado.',
+  },
+  {
+    name: '.ods-select-autocomplete__arrow--disabled',
+    description: 'Estilos aplicados à seta quando disabled=true.',
+  },
+];
+
+// Factory para criar Common Patterns para Select
+export const createSelectCommonPatterns = (): CommonPatternsConfig => ({
+  patterns: [
+    {
+      code: `// Uso básico
+<Select
+  label="País"
+  placeholder="Selecione um país..."
+  helperText="Escolha seu país de residência"
+  options={[
+    { value: 'br', label: 'Brasil' },
+    { value: 'us', label: 'Estados Unidos' },
+    { value: 'fr', label: 'França' }
+  ]}
+/>
+
+// Com validação
+<Select
+  label="Categoria"
+  error={hasError}
+  helperText={hasError ? "Este campo é obrigatório" : "Selecione uma categoria"}
+  options={categories}
+/>
+
+// Em formulários
+<form onSubmit={handleSubmit}>
+  <Select
+    name="priority"
+    label="Prioridade"
+    required
+    options={priorities}
+  />
+</form>`,
+    },
+  ],
+});
+
+// Factory para criar Common Patterns para SelectAutocomplete
+export const createSelectAutocompleteCommonPatterns =
+  (): CommonPatternsConfig => ({
+    patterns: [
+      {
+        code: `// Busca em lista grande
+<SelectAutocomplete
+  label="Cidade"
+  placeholder="Digite para buscar..."
+  helperText="Comece digitando o nome da cidade"
+  options={cities}
+/>
+
+// Com validação de busca
+<SelectAutocomplete
+  label="Produto"
+  error={hasError}
+  helperText={hasError ? "Produto não encontrado" : "Digite para filtrar produtos"}
+  options={products}
+  onInputChange={handleSearch}
+/>
+
+// Em formulários complexos
+<form onSubmit={handleSubmit}>
+  <SelectAutocomplete
+    name="customer"
+    label="Cliente"
+    required
+    options={customers}
+    onChange={handleCustomerSelect}
+  />
+</form>`,
+      },
+    ],
+  });
+
+// Factory para criar Usage Examples para Select
+export const createSelectUsageExamples = (): React.FC => {
+  const SelectUsageExamples = (): JSX.Element => (
+    <>
+      <DocBlock.Heading>Exemplos de Uso</DocBlock.Heading>
+
+      <h3>Seleção de País</h3>
+      <DocBlock.Source
+        dark
+        code={`<Select
+  label="País"
+  placeholder="Selecione seu país..."
+  helperText="Escolha seu país de residência"
+  options={countries}
+  onChange={handleCountryChange}
+/>`}
+      />
+
+      <h3>Formulários com Validação</h3>
+      <DocBlock.Source
+        dark
+        code={`<form onSubmit={handleSubmit}>
+  <Select
+    name="category"
+    label="Categoria"
+    error={hasError}
+    helperText={hasError ? "Categoria é obrigatória" : "Selecione uma categoria"}
+    options={categories}
+    required
+  />
+  <Button type="submit">Salvar</Button>
+</form>`}
+      />
+
+      <h3>Com Estado Controlado</h3>
+      <DocBlock.Source
+        dark
+        code={`const [priority, setPriority] = useState('');
+const [error, setError] = useState(false);
+
+const handlePriorityChange = (value) => {
+  setPriority(value);
+  setError(!value); // Valida se foi selecionado
+};
+
+<Select
+  label="Prioridade"
+  value={priority}
+  onChange={handlePriorityChange}
+  error={error}
+  helperText={error ? "Selecione uma prioridade" : "Defina o nível de urgência"}
+  options={priorities}
+/>`}
+      />
+    </>
+  );
+  return SelectUsageExamples;
+};
+
+// Factory para criar Usage Examples para SelectAutocomplete
+export const createSelectAutocompleteUsageExamples = (): React.FC => {
+  const SelectAutocompleteUsageExamples = (): JSX.Element => (
+    <>
+      <DocBlock.Heading>Exemplos de Uso</DocBlock.Heading>
+
+      <h3>Busca de Cidades</h3>
+      <DocBlock.Source
+        dark
+        code={`<SelectAutocomplete
+  label="Cidade"
+  placeholder="Digite para buscar cidades..."
+  helperText="Comece digitando o nome da cidade"
+  options={cities}
+  onChange={handleCitySelect}
+/>`}
+      />
+
+      <h3>Com Estado Controlado</h3>
+      <DocBlock.Source
+        dark
+        code={`const [inputValue, setInputValue] = useState('');
+const [selectedValue, setSelectedValue] = useState('');
+
+<SelectAutocomplete
+  label="Produto"
+  inputValue={inputValue}
+  value={selectedValue}
+  onInputChange={setInputValue}
+  onChange={setSelectedValue}
+  options={filteredProducts}
+  placeholder="Buscar produtos..."
+/>`}
+      />
+
+      <h3>Com Busca no Servidor</h3>
+      <DocBlock.Source
+        dark
+        code={`const [options, setOptions] = useState([]);
+const [loading, setLoading] = useState(false);
+
+const handleSearch = useMemo(
+  () => debounce(async (query) => {
+    if (query.length < 2) return;
+    setLoading(true);
+    const results = await searchAPI(query);
+    setOptions(results);
+    setLoading(false);
+  }, 300),
+  []
+);
+
+<SelectAutocomplete
+  label="Cliente"
+  placeholder="Digite pelo menos 2 caracteres..."
+  onInputChange={handleSearch}
+  options={options}
+  loading={loading}
+/>`}
+      />
+    </>
+  );
+  return SelectAutocompleteUsageExamples;
+};
+
+// Factory para criar demonstração de tamanhos de lista
+export const createListSizesDemo = (
+  Component: React.ComponentType<any>
+): React.FC => {
+  const ListSizesDemoComponent = (): JSX.Element => (
+    <div style={sharedContainerStyles.showcase}>
+      <Component
+        label="Lista Pequena (5 opções)"
+        placeholder="Selecione..."
+        helperText="Ideal para poucas opções"
+        options={SELECT_COMMON_DATA.PRIORITIES}
+      />
+      <Component
+        label="Lista Média (8 opções)"
+        placeholder="Selecione..."
+        helperText="Boa para categorias principais"
+        options={SELECT_COMMON_DATA.COUNTRIES}
+      />
+      <Component
+        label="Lista Grande (10+ opções)"
+        placeholder="Selecione..."
+        helperText="Para muitas opções, considere busca"
+        options={generateNumberOptions(15, 'Opção')}
+      />
+    </div>
+  );
+  return ListSizesDemoComponent;
+};
+
+// Factory para criar demonstração de casos de uso para SelectAutocomplete
+export const createUseCasesDemo = (
+  Component: React.ComponentType<any>
+): React.FC => {
+  const UseCasesDemoComponent = (): JSX.Element => (
+    <div style={sharedContainerStyles.showcase}>
+      <Component
+        label="Cidade"
+        placeholder="Digite para buscar cidades..."
+        helperText="Comece digitando o nome da cidade"
+        options={LARGE_CITIES}
+      />
+      <Component
+        label="Produto"
+        placeholder="Buscar produtos..."
+        helperText="Digite o nome ou categoria do produto"
+        options={PRODUCTS}
+      />
+      <Component
+        label="Cliente"
+        placeholder="Nome do cliente..."
+        helperText="Digite para filtrar clientes"
+        options={SELECT_COMMON_DATA.COUNTRIES}
+      />
+    </div>
+  );
+  return UseCasesDemoComponent;
+};
+
+// Factory para criar demonstração de performance para SelectAutocomplete
+export const createPerformanceDemo = (
+  Component: React.ComponentType<any>
+): React.FC => {
+  const PerformanceDemoComponent = (): JSX.Element => (
+    <div style={sharedContainerStyles.showcase}>
+      <Component
+        label="Lista Pequena (8 opções)"
+        placeholder="Buscar..."
+        helperText="Use Select simples para poucas opções"
+        options={SELECT_COMMON_DATA.COUNTRIES}
+      />
+      <Component
+        label="Lista Grande (30+ opções)"
+        placeholder="Digite para filtrar..."
+        helperText="Ideal para listas grandes"
+        options={LARGE_CITIES}
+      />
+      <Component
+        label="Lista com Produtos (16 opções)"
+        placeholder="Buscar produtos..."
+        helperText="Ótimo para catálogos"
+        options={PRODUCTS}
+      />
+    </div>
+  );
+  return PerformanceDemoComponent;
+};
+
+// Listas de dados para documentação
+export const SELECT_DATA_TYPES_LIST = [
+  {
+    code: 'Países',
+    description: 'Para seleção de localização e nacionalidade',
+  },
+  {
+    code: 'Categorias',
+    description: 'Para classificação de produtos ou conteúdo',
+  },
+  { code: 'Prioridades', description: 'Para tarefas, tickets e urgência' },
+  { code: 'Status', description: 'Para estados de processo e workflow' },
+];
+
+export const SELECT_STATES_DESCRIPTION = [
+  {
+    state: 'error',
+    description: 'Quando há problemas de validação ou seleção obrigatória',
+  },
+  { state: 'disabled', description: 'Quando a seleção não está disponível' },
+  {
+    state: 'helperText',
+    description: 'Para fornecer contexto sobre as opções',
+  },
+];
+
+export const SELECT_AUTOCOMPLETE_USE_CASES_LIST = [
+  {
+    code: 'Busca de Cidades',
+    description: 'Para listas grandes de localizações geográficas',
+  },
+  {
+    code: 'Seleção de Clientes',
+    description: 'Em sistemas CRM com muitos clientes',
+  },
+  {
+    code: 'Catálogo de Produtos',
+    description: 'Para e-commerce com inventário extenso',
+  },
+  {
+    code: 'Tags e Categorias',
+    description: 'Sistemas de classificação flexíveis',
+  },
+];
+
+export const SELECT_AUTOCOMPLETE_STATES_DESCRIPTION = [
+  {
+    state: 'error',
+    description: 'Quando a busca não retorna resultados válidos',
+  },
+  { state: 'disabled', description: 'Quando a busca não está disponível' },
+  { state: 'loading', description: 'Durante buscas em APIs externas' },
+];
+
 // Factory para criar demonstração de estados
 export const createStatesDemo = (
   Component: React.ComponentType<any>,
@@ -604,6 +966,216 @@ export const createUsageStory = (
   },
   decorators: createDefaultDecorator(),
 });
+
+// Props específicas do Select
+export const SELECT_SPECIFIC_PROPS: ApiProp[] = [
+  {
+    name: 'options',
+    type: 'Array<{ value: string; label: string; disabled?: boolean }>',
+    defaultValue: '[]',
+    description:
+      'Array de opções para seleção. Cada opção deve ter value e label.',
+  },
+  {
+    name: 'value',
+    type: 'string',
+    defaultValue: 'undefined',
+    description: 'O valor atual selecionado (modo controlado).',
+  },
+  {
+    name: 'onChange',
+    type: '(value: string) => void',
+    defaultValue: 'undefined',
+    description: 'Callback chamado quando a seleção muda.',
+  },
+  ...COMMON_SELECT_PROPS,
+];
+
+// Props específicas do SelectAutocomplete
+export const SELECT_AUTOCOMPLETE_SPECIFIC_PROPS: ApiProp[] = [
+  {
+    name: 'options',
+    type: 'Array<{ value: string; label: string; disabled?: boolean }>',
+    defaultValue: '[]',
+    description:
+      'Array de opções para seleção e busca. Cada opção deve ter value e label.',
+  },
+  {
+    name: 'value',
+    type: 'string',
+    defaultValue: 'undefined',
+    description: 'O valor atual selecionado (modo controlado).',
+  },
+  {
+    name: 'onChange',
+    type: '(value: string) => void',
+    defaultValue: 'undefined',
+    description: 'Callback chamado quando a seleção muda.',
+  },
+  {
+    name: 'onInputChange',
+    type: '(inputValue: string) => void',
+    defaultValue: 'undefined',
+    description: 'Callback chamado quando o texto de busca muda.',
+  },
+  {
+    name: 'inputValue',
+    type: 'string',
+    defaultValue: 'undefined',
+    description: 'Valor atual do campo de busca (modo controlado).',
+  },
+  {
+    name: 'noOptionsText',
+    type: 'string',
+    defaultValue: '"Nenhuma opção"',
+    description: 'Texto exibido quando nenhuma opção corresponde à busca.',
+  },
+  ...COMMON_SELECT_PROPS,
+];
+
+// Configurações compartilhadas
+export const SELECT_INTRODUCTION_CONFIG: IntroductionConfig = {
+  description:
+    'Componente de seleção que permite ao usuário escolher uma opção de uma lista suspensa.',
+  overview:
+    'O componente Select oferece uma interface limpa e intuitiva para seleção de opções em listas. É ideal para formulários onde o usuário precisa escolher entre múltiplas alternativas pré-definidas. Suporta placeholder, validação de erro, estado desabilitado e texto de ajuda.',
+  characteristics: [
+    {
+      title: 'Lista Suspensa',
+      description: 'Interface compacta que se expande para mostrar opções',
+    },
+    {
+      title: 'Busca Integrada',
+      description: 'Permite filtrar opções conforme o usuário digita',
+    },
+    {
+      title: 'Validação',
+      description: 'Estados de erro e validação em tempo real',
+    },
+    {
+      title: 'Acessibilidade',
+      description: 'Navegação por teclado e suporte a leitores de tela',
+    },
+    {
+      title: 'Customizável',
+      description: 'Suporte a placeholder, helper text e estados visuais',
+    },
+  ],
+  importPath: `import { Select } from '@useblu/ocean-react';`,
+};
+
+export const SELECT_AUTOCOMPLETE_INTRODUCTION_CONFIG: IntroductionConfig = {
+  description:
+    'Componente de seleção com busca integrada que permite ao usuário filtrar opções digitando.',
+  overview:
+    'O SelectAutocomplete combina a funcionalidade de um campo de texto com uma lista suspensa, permitindo que o usuário digite para filtrar opções em tempo real. É ideal para listas grandes onde a busca é essencial para encontrar rapidamente a opção desejada.',
+  characteristics: [
+    {
+      title: 'Busca em Tempo Real',
+      description: 'Filtra opções conforme o usuário digita',
+    },
+    {
+      title: 'Performance Otimizada',
+      description: 'Otimizado para listas grandes com centenas de opções',
+    },
+    {
+      title: 'Flexibilidade',
+      description: 'Permite entrada personalizada ou apenas seleção de opções',
+    },
+    {
+      title: 'Navegação por Teclado',
+      description: 'Suporte completo a navegação por setas e Enter',
+    },
+    {
+      title: 'Acessibilidade',
+      description: 'Compatível com leitores de tela e ARIA',
+    },
+  ],
+  importPath: `import { SelectAutocomplete } from '@useblu/ocean-react';`,
+};
+
+// Configurações de Best Practices comuns
+export const SELECT_BEST_PRACTICES_CONFIG: BestPracticesConfig = {
+  sections: [
+    {
+      title: '1. Uso Geral',
+      items: [
+        'Use para listas de 3 a 100 opções para melhor performance',
+        'Forneça labels descritivos que indiquem claramente o propósito',
+        'Use placeholders que orientem sobre a ação esperada',
+        'Mantenha as opções em ordem lógica (alfabética, por frequência)',
+      ],
+    },
+    {
+      title: '2. Organização das Opções',
+      items: [
+        'Para até 10 opções: ordem alfabética ou por importância',
+        'Para 10-50 opções: considere agrupamento por categoria',
+        'Para 50+ opções: implemente busca ou use SelectAutocomplete',
+        'Coloque opções mais comuns no topo quando apropriado',
+      ],
+    },
+    {
+      title: '3. Acessibilidade',
+      items: [
+        'Sempre forneça labels apropriados para leitores de tela',
+        'Use helperText para orientações específicas',
+        'Mantenha navegação por teclado funcional (Enter, Esc, setas)',
+        'Forneça feedback claro sobre seleção atual',
+      ],
+    },
+    {
+      title: '4. Design e UX',
+      items: [
+        'Use placeholders que indiquem a ação (&ldquo;Selecione...&rdquo;)',
+        'Forneça feedback visual claro para estados de erro',
+        'Considere o contexto mobile - opções grandes o suficiente',
+        'Mantenha consistência com outros campos do formulário',
+      ],
+    },
+  ],
+};
+
+export const SELECT_AUTOCOMPLETE_BEST_PRACTICES_CONFIG: BestPracticesConfig = {
+  sections: [
+    {
+      title: '1. Quando Usar',
+      items: [
+        'Para listas com mais de 50 opções onde busca é essencial',
+        'Quando usuários precisam encontrar rapidamente opções específicas',
+        'Para dados como cidades, clientes, produtos com nomes únicos',
+        'Substitua Select comum quando usuários reportam dificuldade para encontrar opções',
+      ],
+    },
+    {
+      title: '2. Performance e Dados',
+      items: [
+        'Use para listas de até 1000 opções com boa performance',
+        'Para mais de 1000 itens, considere paginação ou busca no servidor',
+        'Mantenha dados ordenados alfabeticamente para melhor UX',
+        'Use debounce para buscas em APIs externas',
+      ],
+    },
+    {
+      title: '3. UX e Busca',
+      items: [
+        'Forneça placeholders que orientem sobre a busca (&ldquo;Digite para buscar...&rdquo;)',
+        'Use helperText para explicar como a busca funciona',
+        'Considere mostrar &ldquo;Nenhum resultado&rdquo; quando a busca não retorna opções',
+        'Permita limpar a busca facilmente',
+      ],
+    },
+    {
+      title: '4. Acessibilidade',
+      items: [
+        'Mantenha foco visível durante navegação por teclado',
+        'Anuncie resultados de busca para leitores de tela',
+        'Use aria-labels apropriados para estado da busca',
+        'Garanta que Escape limpe a busca e feche a lista',
+      ],
+    },
+  ],
+};
 
 // Template para página de documentação
 export const createDocsPage = (config: {
